Remove Teleios from home page and create dedicated page
